feature:削除確認モーダルウィンのコンポーネント化#15
