+++ backend/app/controllers/api/v1/auth/registrations_controller.rb
@@ -1,5 +1,14 @@
 class Api::V1::Auth::RegistrationsController < DeviseTokenAuth::RegistrationsController
 
+  def create
+    user = User.new(sign_up_params)
+    if user.save
+      render json: user
+    else
+      render json: { errors: user.errors.full_messages }, status: 422
+    end
+  end
+
   # def destroy
   #   user = User.find_by(params[:id])
   #   user.destroy!
