+++ frontend/pages/users/_id.vue
@@ -1,22 +1,6 @@
 <template>
   <v-container>
-    <v-dialog v-model="dialog.isDisplay" width="500">
-      <v-card>
-        <v-card-title class="text-h5 grey lighten-2">
-          本当に削除しますか？
-        </v-card-title>
-        <v-divider></v-divider>
-        <v-card-actions>
-          <v-spacer></v-spacer>
-          <v-btn color="#f06966" text @click="deleteUser">
-            削除する
-          </v-btn>
-          <v-btn color="#6abe83" text @click="dialog.isDisplay = false">
-            戻る
-          </v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-dialog>
+    <UserDeleteModal :dialog="dialog" />
     <v-card width="800px" class="mx-auto mt-5" elevation="1">
       <v-card-title>
         <h1 class="headline">
@@ -58,7 +42,6 @@
 <script lang="ts">
 import {
   defineComponent,
-  useContext,
   inject,
   useFetch,
   reactive,
@@ -68,9 +51,7 @@ import { UseUser } from '@/store/user/userTypes'
 
 export default defineComponent({
   setup() {
-    const { $axios, $auth } = useContext()
-
-    const { user, fetchUser, unsetUser } = inject(userKey) as UseUser
+    const { user, fetchUser } = inject(userKey) as UseUser
 
     const dialog = reactive({
       isDisplay: false,
@@ -80,35 +61,9 @@ export default defineComponent({
       await fetchUser()
     })
 
-    const deleteUser = async () => {
-      await $axios
-        .$delete('/api/v1/auth', {
-          headers: {
-            'access-token': localStorage.getItem('access-token'),
-            client: localStorage.getItem('client'),
-            uid: localStorage.getItem('uid'),
-            'token-type': localStorage.getItem('token-type'),
-          },
-        })
-        .catch((e) => console.log(e))
-      await unsetUser()
-      await $auth
-        .logout()
-        .then(() => {
-          localStorage.removeItem('access-token')
-          localStorage.removeItem('client')
-          localStorage.removeItem('uid')
-          localStorage.removeItem('token-type')
-        })
-        .catch((e) => {
-          console.log(e)
-        })
-    }
-
     return {
       user,
       dialog,
-      deleteUser,
     }
   },
 })
